+++ beecomb-common-backend/src/main/java/io/github/icodegarden/beecomb/common/backend/pojo/transfer/CreateDelayJobDTO.java
@@ -18,7 +18,7 @@ import lombok.ToString;
 public class CreateDelayJobDTO implements Validateable {
 
 	private Long jobId;
-	private Integer delay;// int comment 'ms',
+	private Long delay;// int comment 'ms',
 	private Integer retryOnExecuteFailed = 0;// smallint NOT NULL DEFAULT 0 comment 'executor执行失败重试次数，包括连接失败、超时等',
 	private Integer retryBackoffOnExecuteFailed = 3000;// int NOT NULL DEFAULT 3000 comment 'ms要求 gte 1000',
 	private Integer retryOnNoQualified = 0;// smallint NOT NULL DEFAULT 0 comment '没有可用的executor时重试次数，包括不在线、超载时',
